lightnvm: set correct read size when reading log pages
Number of words to read when issuing a get log page request
is a zero based value, so correct the request setup.

From the NVM Express 1.3 specification, figure 86

"The combined NUMDL and NUMDU fields form a 0’s based value"

This is also how the nvme cli tool sets up the log page requests.

Signed-off-by: Hans Holmberg <hans.holmberg@cnexlabs.com>