lightnvn: pblk: implement get log report chunk
On the OCCSSD 2.0 spec, bad block discovery is done through the nvme get
log page command. Implement this functionality on pblk for the 2.0 path.

Also, since there is no need to mark bad blocks on the 2.0 path in case
of write/erase errors, only mark the internal bitmaps (and avoid
sending the mark_bb command to the device).

Signed-off-by: Javier González <javier@cnexlabs.com>